+++ src/lib/Components/Bidding/Components/Markdown.tsx
@@ -1,3 +1,4 @@
+import { Serif } from "@artsy/palette"
 import _ from "lodash"
 import React from "react"
 import { Text, ViewProperties } from "react-native"
@@ -5,7 +6,6 @@ import SimpleMarkdown from "simple-markdown"
 
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { Flex, FlexProps } from "../Elements/Flex"
-import { Serif16t } from "../Elements/Typography"
 import { LinkText } from "./LinkText"
 
 // Rules for rendering parsed markdown. Currently only handles links and text. Add rules similar to
@@ -44,9 +44,9 @@ const rules = {
     match: SimpleMarkdown.blockRegex(/^((?:[^\n]|\n(?! *\n))+)(?:\n *)/),
     react: (node, output, state) => {
       return (
-        <Serif16t color="black60" key={state.key} textAlign="center">
+        <Serif size="3t" color="black60" key={state.key} textAlign="center">
           {output(node.content, state)}
-        </Serif16t>
+        </Serif>
       )
     },
   },